Have JSON numbers default to double instead of int to keep precision
Summary:
Previously, all numbers were cast to longs and therefore
any precision past the decimal point was lost. This fixes that.
(closes #3913)

Test Plan: Loaded in Tableau and verified

Reviewers: kw, sandeep, purvi

Reviewed By: kw, sandeep

Differential Revision: https://rockset.phacility.com/D3509

+++ src/main/java/com/rockset/jdbc/RocksetResultSet.java
@@ -1454,7 +1454,7 @@ public class RocksetResultSet implements ResultSet {
       return 0;
     }
     if (value instanceof NumericNode) {
-      return (Number) (((NumericNode)value).longValue());
+      return (Number) (((NumericNode)value).doubleValue());
     }
     if (value instanceof BooleanNode) {
       BooleanNode n = (BooleanNode)(value);
